refactor: start -> startGame, startNewRound -> startRound
