+++ src/backend/game/GameMethods/common.js
@@ -1,3 +1,7 @@
+import { PLAYER, CARD, TIME } from '@utils/number';
+import GAME_STATE from '@utils/gameState';
+import { emit } from '@socket';
+
 function getState() {
   return this.status.state;
 }
@@ -10,6 +14,32 @@ function setState(state) {
   this.status.state = state;
 }
 
-const methodGroup = { getState, setState };
+function startRound() {
+  this.addTurn();
+  this.initFirstDiscussionUser();
+
+  const {
+    users,
+    status: { unusedCards, turn },
+  } = this;
+
+  const isFirstTurn = turn === 1;
+  const teller = [...users.values()][turn % users.size];
+  const { socketID: tellerID } = teller;
+  const requiredCardCount = isFirstTurn ? CARD.HAND : 1;
+  const outOfDeck = unusedCards.length < users.size * requiredCardCount;
+  if (outOfDeck) return;
+
+  users.forEach((user) => {
+    const cards = this.dealCards(user.cards, requiredCardCount);
+    const params = { tellerID, cards, endTime: this.endTime };
+    user.initOnRound(params);
+    emit({ socketID: user.socketID, name: 'get round data', params });
+  });
+
+  this.waitTellerSelect(tellerID);
+}
+
+const methodGroup = { getState, setState, startRound };
 
 export default methodGroup;
